user-6, general cleanup of main user README.md
